requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.12.0,<3.0.0
pyyaml==6.0.2
pyperclip==1.9.0
openai==2.21.0
google-genai>=1.0.0,<2.0.0
jinja2==3.1.6
python-multipart==0.0.22
aiofiles==24.1.0
packaging==24.2
pywebview==5.4
pystray==0.19.5
pillow==12.1.1
websockets>=12.0,<14.0
httpx>=0.27.0,<1.0.0
